feat: 음소거 기능 완료 및 voicerecorder컴포넌트 정리
